feat:Todo 앱 구현, 컴포넌트 분리
